source/cell-container/requirements.txt
Flask
boto3
pyjwt
Flask-HTTPAuth